refactor : 추가요금계산을 위한 OverFareCalculator 생성 및 거리에 따른 추가요금 계산 로직 이동
