+++ docs/subway.md
@@ -75,6 +75,7 @@
   - [X] 이동거리 기반 요금 산정 기능 추가
   - [ ] 노선별 요금 정책 반영
     - [X] 최단경로 검색결과 객체에 노선정보 추가
+    - [X] 추가요금계산을 위한 OverFareCalculator 생성 및 거리에 따른 추가요금 계산 로직 이동
   - [ ] 연령별 요금 정책 반영
 
 ---
