+++ src/test/java/nextstep/subway/path/domain/SubwayFareTest.java
@@ -2,7 +2,6 @@ package nextstep.subway.path.domain;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static org.assertj.core.api.Assertions.*;
@@ -43,12 +42,4 @@ class SubwayFareTest {
         SubwayFare fare = SubwayFare.of(10000);
         assertThatThrownBy(() ->  fare.discountedByPercent(discount)).isInstanceOf(IllegalArgumentException.class);
     }
-
-    @ParameterizedTest
-    @CsvSource(delimiterString = ":", value ={"5:1250", "10:1250","11:1350","20:1450","50:2050","51:2150","66:2250","67:2350"} )
-    void 거리에_따른_지하철요금_계산(String distance, String expectedFare){
-        SubwayFare fare = SubwayFare.calculateByDistance(Integer.parseInt(distance));
-        assertThat(fare).isEqualTo(SubwayFare.of(Integer.parseInt(expectedFare)));
-    }
-
 }
